requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
openai==1.5.0
python-dotenv==1.0.0
retrying==1.3.4